update: name changes
In horizontal mode, aggregator is now split into top aggregator and middle aggregator. To adapt this change, this PR addressed all the name changes related to this naming convention, and made sure the previous examples still work.
