+++ lib/python/fledge/examples/hier_mnist/middle_aggregator/__init__.py
