+++ lib/python/fledge/examples/hier_mnist/middle_aggregator/config_uk.json
