+++ lib/python/fledge/examples/hier_mnist/top_aggregator/config.json
