+++ lib/python/fledge/mode/horizontal/top_aggregator.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2022 Cisco Systems, Inc. and its affiliates
+# Copyright (c) 2021 Cisco Systems, Inc. and its affiliates
 # All rights reserved
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-""" honrizontal FL middle level aggregator"""
+"""horizontal FL top level aggregator"""
 
 import logging
 import time
@@ -25,6 +25,7 @@ from ...common.custom_abcmeta import ABCMeta, abstract_attribute
 from ...optimizer.train_result import TrainResult
 from ...optimizers import optimizer_provider
 from ...plugin import PluginManager, PluginType
+from ...registries import registry_provider
 from ..composer import Composer
 from ..message import MessageType
 from ..role import Role
@@ -34,11 +35,10 @@ logger = logging.getLogger(__name__)
 
 TAG_DISTRIBUTE = 'distribute'
 TAG_AGGREGATE = 'aggregate'
-TAG_FETCH = 'fetch'
-TAG_UPLOAD = 'upload'
 
-class MiddleAggregator(Role, metaclass=ABCMeta):
-    """ Middle level aggregator acts as both top level aggregator and trainer."""
+
+class TopAggregator(Role, metaclass=ABCMeta):
+    """Top level Aggregator implements an ML aggregation role."""
 
     @abstract_attribute
     def weights(self):
@@ -48,6 +48,7 @@ class MiddleAggregator(Role, metaclass=ABCMeta):
     def config(self):
         """Abstract attribute for config object."""
 
+    @abstract_attribute
     def metrics(self):
         """
         Abstract attribute for metrics object.
@@ -67,10 +68,6 @@ class MiddleAggregator(Role, metaclass=ABCMeta):
         dataset's type is Dataset (in fledge/dataset.py).
         """
 
-    @abstract_attribute
-    def dataset_size(self):
-        """Abstract attribute for size of dataset used to train."""
-
     def internal_init(self) -> None:
         """Initialize internal state for role."""
         # global variable for plugin manager
@@ -80,61 +77,31 @@ class MiddleAggregator(Role, metaclass=ABCMeta):
         self.cm(self.config)
         self.cm.join_all()
 
+        self.registry_client = registry_provider.get(self.config.registry.sort)
+        # initialize registry client
+        self.registry_client(self.config.registry.uri, self.config.job.job_id)
+
+        base_model = self.config.base_model
+        if base_model and base_model.name != "" and base_model.version > 0:
+            self.model = self.registry_client.load_model(
+                base_model.name, base_model.version)
+
+        self.registry_client.setup_run()
+
+        # disk cache is used for saving memory in case model is large
+        self.cache = Cache()
         self.optimizer = optimizer_provider.get(self.config.optimizer)
 
+        self._round = 1
+        self._rounds = 1
+        if 'rounds' in self.config.hyperparameters:
+            self._rounds = self.config.hyperparameters['rounds']
         self._work_done = False
-
-        self.cache = Cache()
 
     def get(self, tag: str) -> None:
         """Get data from remote role(s)."""
-        if tag == TAG_FETCH:
-            self._fetch_weights(tag)
         if tag == TAG_AGGREGATE:
             self._aggregate_weights(tag)
-
-    def put(self, tag: str) -> None:
-        """Set data to remote role(s)."""
-        if tag == TAG_UPLOAD:
-            self._send_weights(tag)
-        if tag == TAG_DISTRIBUTE:
-            self.dist_tag = tag
-            self._distribute_weights(tag)
-
-    def _fetch_weights(self, tag: str) -> None:
-        logger.debug("calling _fetch_weights")
-        channel = self.cm.get_by_tag(tag)
-        if not channel:
-            logger.debug(f"[_fetch_weights] channel not found with tag {tag}")
-            return
-
-        while channel.empty():
-            time.sleep(1)
-            logger.debug("[_fetch_weights] waiting for channel ends")
-
-        # one aggregator is sufficient
-        end = channel.one_end()
-        dict = channel.recv(end)
-        for k, v in dict.items():
-            if k == MessageType.WEIGHTS:
-                self.weights = v
-            elif k == MessageType.EOT:
-                self._work_done = v
-
-    def _distribute_weights(self, tag: str) -> None:
-        channel = self.cm.get_by_tag(tag)
-        if not channel:
-            logger.debug(f"channel not found for tag {tag}")
-            return
-
-        while channel.empty():
-            logger.debug("no end found in the channel")
-            time.sleep(1)
-
-        # send out global model parameters to trainers
-        for end in channel.ends():
-            logger.debug(f"sending weights to {end}")
-            channel.send(end, {MessageType.WEIGHTS: self.weights})
 
     def _aggregate_weights(self, tag: str) -> None:
         channel = self.cm.get_by_tag(tag)
@@ -168,25 +135,27 @@ class MiddleAggregator(Role, metaclass=ABCMeta):
 
         # set global weights
         self.weights = global_weights
-        self.dataset_size += total
 
-    def _send_weights(self, tag: str) -> None:
-        logger.debug("calling _send_weights")
+    def put(self, tag: str) -> None:
+        """Set data to remote role(s)."""
+        if tag == TAG_DISTRIBUTE:
+            self.dist_tag = tag
+            self._distribute_weights(tag)
+
+    def _distribute_weights(self, tag: str) -> None:
         channel = self.cm.get_by_tag(tag)
         if not channel:
-            logger.debug(f"[_send_weights] channel not found with {tag}")
+            logger.debug(f"channel not found for tag {tag}")
             return
 
         while channel.empty():
+            logger.debug("no end found in the channel")
             time.sleep(1)
-            logger.debug("[_send_weights] waiting for channel ends")
 
-        # one aggregator is sufficient
-        end = channel.one_end()
-
-        data = (self.weights, self.dataset_size)
-        channel.send(end, data)
-        logger.debug("sending weights done")
+        # send out global model parameters to trainers
+        for end in channel.ends():
+            logger.debug(f"sending weights to {end}")
+            channel.send(end, {MessageType.WEIGHTS: self.weights})
 
     def inform_end_of_training(self) -> None:
         """Inform all the trainers that the training is finished."""
@@ -212,6 +181,38 @@ class MiddleAggregator(Role, metaclass=ABCMeta):
             # merge metrics with the existing metrics
             self.metrics = self.metrics | metrics
 
+    def save_metrics(self):
+        """Save metrics in a model registry."""
+        logger.debug(f"saving metrics: {self.metrics}")
+        if self.metrics:
+            self.registry_client.save_metrics(self._round - 1, self.metrics)
+            logger.debug("saving metrics done")
+
+    def increment_round(self):
+        """Increment the round counter."""
+        logger.debug(f"Incrementing current round: {self._round}")
+        self._round += 1
+        self._work_done = (self._round > self._rounds)
+
+        channel = self.cm.get_by_tag(self.dist_tag)
+        if not channel:
+            logger.debug(f"channel not found for tag {self.dist_tag}")
+            return
+
+        # set necessary properties to help channel decide how to select ends
+        channel.set_property("round", self._round)
+
+    def save_params(self):
+        """Save hyperparamets in a model registry."""
+        if self.config.hyperparameters:
+            self.registry_client.save_params(self.config.hyperparameters)
+
+    def save_model(self):
+        """Save model in a model registry."""
+        if self.model:
+            model_name = f"{self.config.job.name}-{self.config.job.job_id}"
+            self.registry_client.save_model(model_name, self.model)
+
     def compose(self) -> None:
         """Compose role with tasklets."""
         with Composer() as composer:
@@ -223,31 +224,38 @@ class MiddleAggregator(Role, metaclass=ABCMeta):
 
             task_load_data = Tasklet(self.load_data)
 
-            task_put_dist = Tasklet(self.put, TAG_DISTRIBUTE)
+            task_put = Tasklet(self.put, TAG_DISTRIBUTE)
 
-            task_put_upload = Tasklet(self.put, TAG_UPLOAD)
+            task_get = Tasklet(self.get, TAG_AGGREGATE)
 
-            task_get_aggr = Tasklet(self.get, TAG_AGGREGATE)
-
-            task_get_fetch = Tasklet(self.get, TAG_FETCH)
+            task_train = Tasklet(self.train)
 
             task_eval = Tasklet(self.evaluate)
 
             task_analysis = Tasklet(self.run_analysis)
 
+            task_save_metrics = Tasklet(self.save_metrics)
+
+            task_increment_round = Tasklet(self.increment_round)
+
             task_end_of_training = Tasklet(self.inform_end_of_training)
+
+            task_save_params = Tasklet(self.save_params)
+
+            task_save_model = Tasklet(self.save_model)
 
         # create a loop object with loop exit condition function
         loop = Loop(loop_check_fn=lambda: self._work_done)
         task_internal_init >> task_init >> loop(
-            task_load_data >> task_get_fetch >> task_put_dist >> task_get_aggr >> task_put_upload 
-            >> task_eval >> task_analysis 
-        ) >> task_end_of_training 
+            task_load_data >> task_put >> task_get >> task_train >> task_eval
+            >> task_analysis >> task_save_metrics >> task_increment_round
+        ) >> task_end_of_training >> task_save_params >> task_save_model
 
     def run(self) -> None:
         """Run role."""
         self.composer.run()
 
+    @classmethod
     def get_func_tags(cls) -> list[str]:
-        """Return a list of function tags defined in the middle level aggregator role."""
-        return [TAG_DISTRIBUTE, TAG_AGGREGATE, TAG_FETCH, TAG_UPLOAD]
+        """Return a list of function tags defined in the top level aggregator role."""
+        return [TAG_DISTRIBUTE, TAG_AGGREGATE]
